feat: Implement BillData and BillItem models for OCR parsing and bill management
