+++ lib/models/bill_item.dart
@@ -1,0 +1,13 @@
+class BillItem {
+  final String name;
+  final double price;
+  final int quantity;
+  final double total;
+
+  BillItem({
+    required this.name,
+    required this.price,
+    required this.quantity,
+    required this.total,
+  });
+}
